implemented the connect function. The idea is that you connect to an already existing file which will be used for any function run from that instance of JsonTools. Ideally you would have to run the connect function before you do anything else
For example if you have an instance of JsonTools called "j", and you connect to "test.json" any "j." function that you run will use the file you initially connected to.

<======================>
implemented a read to dictionary command, returns the contents of the .json file as a dict.

<======================>
implemented a dump sample command, mainly for debugging. dumps a very simple dictionary to the .json file